Move types parsing to another file
